+++ src/core/index.ts
@@ -1,3 +1,4 @@
 export { createUnistylesComponent } from './createUnistylesComponent'
 export { createUnistylesElement } from './createUnistylesElement'
 export { createUnistylesImageBackground } from './createUnistylesImageBackground'
+export { withUnistyles } from './withUnistyles'
