+++ src/core/withUnistyles/index.ts
@@ -1,0 +1,1 @@
+export { withUnistyles } from './withUnistyles'
