+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from './specs'
 export { mq } from './mq'
 export type { UnistylesThemes, UnistylesBreakpoints } from './global'
-export { createUnistylesComponent } from './core'
+export { createUnistylesComponent, withUnistyles } from './core'
 export type { UnistylesVariants } from './types'
 export { Display, Hide, Variants, ScopedTheme } from './components'
